Спринт 15. Дописал request, внес изменения в booking, остались тесты

+++ src/main/java/ru/practicum/shareit/item/service/ItemServiceImpl.java
@@ -2,6 +2,8 @@ package ru.practicum.shareit.item.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.shareit.booking.model.Booking;
@@ -19,12 +21,10 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 
-@Transactional(readOnly = true)
 @Service
 @RequiredArgsConstructor
 @Slf4j
 public class ItemServiceImpl implements ItemService {
-
 
     private final ItemRepository itemRepository;
     private final BookingRepository bookingRepository;
@@ -38,9 +38,12 @@ public class ItemServiceImpl implements ItemService {
     }
 
     @Override
-    public List<Item> getAllItemsByOwnerId(Long ownerId) {
-        List<Item> itemList = itemRepository.findByOwnerId(ownerId);
-        itemList.sort(Comparator.comparing(Item::getId));
+    public List<Item> getAllItemsByOwnerId(Long ownerId, Integer from, Integer size) {
+        Pageable pageable = PageRequest.of(from / size, size);
+        List<Item> itemList = itemRepository.findByOwnerId(ownerId, pageable);
+        if (itemList.size() > 1) {
+            itemList.sort(Comparator.comparing(Item::getId));
+        }
         return itemList;
     }
 
@@ -51,7 +54,6 @@ public class ItemServiceImpl implements ItemService {
         return itemRepository.save(item);
     }
 
-    @Transactional
     @Override
     public Item updateItem(Long ownerId, Item item) {
         Item updateItem = itemRepository.findById(item.getId())
@@ -71,15 +73,16 @@ public class ItemServiceImpl implements ItemService {
         return itemRepository.save(updateItem);
     }
 
+    @Transactional
     @Override
-    public List<Item> searchItemsByTextInNameAndDescription(String text) {
+    public List<Item> searchItemsByTextInNameAndDescription(String text, Integer from, Integer size) {
         if (text.isBlank()) {
             return Collections.emptyList();
         }
-        return itemRepository.searchItemsByTextInNameAndDescription(text);
+        Pageable pageable = PageRequest.of(from / size, size);
+        return itemRepository.searchItemsByTextInNameAndDescription(text, pageable);
     }
 
-    @Transactional
     @Override
     public Comment createComment(Long userId, Long itemId, Comment comment) {
         Booking booking = bookingRepository.findCompletedBooking(userId, itemId, LocalDateTime.now());
@@ -97,7 +100,16 @@ public class ItemServiceImpl implements ItemService {
     }
 
     @Override
-    public void deleteItem(Long itemId) {
+    public List<Item> getItemsByRequestId(Long requestId) {
+        return itemRepository.findItemsByRequestId(requestId);
+    }
+
+    @Override
+    public void deleteItemById(Long ownerId, Long itemId) {
+        Item item = getItemById(itemId);
+        if (!item.getOwner().getId().equals(ownerId)) {
+            throw new ValidationException("Удалить вещь может только её владелец");
+        }
         itemRepository.deleteById(itemId);
     }
 }
